fix: calculateTotalRevenue
