+++ src/routes/order.route.js
@@ -4,6 +4,7 @@ import orderController from '../controllers/order.controller';
 const router = Express.Router();
 router.get('', orderController.getAll);
 router.get('/analysis', orderController.analysis);
+router.get('/calculateTotalRevenue', orderController.calculateTotalRevenue);
 router.get('/:id', orderController.getOne);
 router.post('', orderController.create);
 router.patch('/:id', orderController.update);
