+++ src/services/company_service.service.js
@@ -1,7 +1,7 @@
 import moment from 'moment/moment';
 import dotenv from 'dotenv';
 import createError from 'http-errors';
-import { Sequelize, company_service, sequelize, service, service_type } from '../models';
+import { Sequelize, company_service, service, service_type } from '../models';
 import { findByPkAndUpdate, findByPkAndDelete } from '../helpers/databaseHelpers';
 
 dotenv.config();
@@ -99,67 +99,6 @@ const companyServiceService = {
 			throw createError(404, 'Không tìm thấy bản ghi');
 		}
 		return results;
-	},
-
-	async calculateTotalRevenue(query) {
-		const { startDate, endDate } = query;
-		const daysBetween = this.calculateDaysDifference(startDate, endDate);
-		const results = await company_service.findAll({
-			attributes: [
-				[sequelize.fn('SUM', sequelize.col('service.price')), 'total_revenue'],
-				[sequelize.col('service.name'), 'service_name'],
-				[sequelize.fn('DATE_FORMAT', sequelize.col('company_service.createdAt'), '%d/%m'), 'day']
-			],
-
-			where: {
-				createdAt: {
-					[Sequelize.Op.between]: [
-						moment(startDate).format('YYYY-MM-DD h:mm:ss'),
-						moment(endDate).format('YYYY-MM-DD h:mm:ss')
-					]
-				}
-			},
-
-			include: [
-				{
-					model: service,
-					attributes: []
-				}
-			],
-
-			group: ['day', 'service.name'],
-			raw: true
-		});
-
-		const label = [];
-
-		const currentDate = new Date(startDate);
-		while (currentDate <= new Date(endDate)) {
-			const formattedDate = `${moment(currentDate).format('DD')}/${currentDate.getMonth() + 1}`;
-			label.push(formattedDate);
-			currentDate.setDate(currentDate.getDate() + 1);
-		}
-		const data = Array(daysBetween).fill(0);
-
-		results.forEach((result) => {
-			const dayIndex = label.indexOf(result.day);
-			data[dayIndex] = Number(result.total_revenue);
-		});
-
-		return {
-			data,
-			label
-		};
-	},
-
-	calculateDaysDifference(start_date, end_date) {
-		const startDateObj = new Date(start_date);
-		const endDateObj = new Date(end_date);
-
-		const timeDifference = endDateObj - startDateObj;
-		const daysDifference = Math.ceil(timeDifference / (24 * 60 * 60 * 1000));
-
-		return daysDifference + 1;
 	}
 };
